Allow for defining SHA1DC_FORCE_UNALIGNED_ACCESS externally
This allows for doing -DSHA1DC_FORCE_UNALIGNED_ACCESS on a
platform/CPU that's not listed here, since the default is to fall back
to aligned access unless except on a whitelist of CPUs.

It would be nice e.g. for the git project if someone running on such a
platform didn't have to monkeypatch the code, instead they can just
provide a -DSHA1DC_FORCE_UNALIGNED_ACCESS argument